+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fix not appearing fixed background images for Chrome 67
 - Fix having a Slider within a Slideshow element
 - Fix filter animation in Edge
+- Fix accessing sessionStorage if that's prevented
 
 ## 3.0.0 rc 4 (May 30, 2018)